requirements.txt
altgraph==0.17.4
colorama==0.4.6
packaging==24.1
pefile==2023.2.7
pillow==10.4.0
pillow_heif==0.18.0
pyinstaller==6.9.0
pyinstaller-hooks-contrib==2024.7
pywin32-ctypes==0.2.2
setuptools==72.1.0
tqdm==4.66.5
